# src/main/resources/swagger/inventory.yml
swagger: "2.0"
info:
  title: "Inventory Microservice"
  description: "API for user accounts."
  version: "1.1.0"
  contact:
    email: "aditya.7sinha@gmail.com"
host: "localhost"
basePath: "/inventory-service/v1"
schemes:
  - http
consumes:
  - application/json;charset=utf-8
produces:
  - application/json;charset=utf-8
tags:
  - name: event
    description: "Everything about an event"
  # - name: events
  #   description: "Everything about list of events"
  - name: issue
    description: "Everything about an issue"
  # - name: issues
  #   description: "Everything about list of issues"
paths:
  /event:
    post:
      tags:
        - event
      summary: "This operation is to create event"
      parameters:
        - in: "body"
          name: "body"
          description: "Event object that needs to be added to the events"
          required: true
          schema:
            $ref: "#/definitions/Event_Create"
      responses:
        "201":
          description: "Event Created"
          schema:
            $ref: "#/definitions/Event"
        "401":
          description: "Invalid"
          schema:
            $ref: "#/definitions/Error"
  /event/{eventid}:
    patch:
      tags:
        - event
      summary: "This operation is to patch event"
      parameters:
        - in: body
          name: "body"
          description: "Event object that needs to be added to the events"
          required: true
          schema:
            $ref: "#/definitions/Event"
        - in: path
          name: "eventid"
          description: "eventId to be updated"
          required: true
          type: string
      responses:
        "201":
          description: "Event Updated"
          schema:
            $ref: "#/definitions/Event"
        "401":
          description: "Invalid"
          schema:
            $ref: "#/definitions/Error"
    get:
      summary: "retrieving Event by Id"
      description: "This operation retrieves a Event entity."
      tags:
        - "event"
      parameters:
        - name: "eventid"
          description: "identification of the user who creates entity"
          required: true
          type: string
          in: path
      responses:
        "201":
          description: "Success"
          schema:
            $ref: "#/definitions/Event"
        "401":
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
  /event/:
    get:
      summary: "retrieving all Events"
      tags:
        - "event"
      responses:
        "201":
          description: "Success"
          schema:
            $ref: "#/definitions/ArrayOfEvents"
        "401":
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
  /event/events/{userid}:
    get:
      summary: "retrieving all Events by UserId"
      tags:
        - "event"
      parameters:
        - name: "userid"
          description: "userid of the user who created the events"
          required: true
          type: string
          in: path
      responses:
        "201":
          description: "Success"
          schema:
            $ref: "#/definitions/ArrayOfEvents"
        "401":
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
  /issue:
    post:
      tags:
        - issue
      summary: "This operation is to create issue"
      parameters:
        - in: "body"
          name: "body"
          description: "Issue object that needs to be added to the issues"
          required: true
          schema:
            $ref: "#/definitions/Issue_Create"
      responses:
        "201":
          description: "Issue Created"
          schema:
            $ref: "#/definitions/Issue"
        "401":
          description: "Invalid"
          schema:
            $ref: "#/definitions/Error"
  /issue/{issueid}:
    patch:
      tags:
        - issue
      summary: "This operation is to patch issue"
      parameters:
        - name: "issueid"
          description: "identification of the user who creates entity"
          required: true
          type: string
          in: path
        - in: "body"
          name: "body"
          description: "Issue object that needs to be added to the issues"
          required: true
          schema:
            $ref: "#/definitions/Issue"
      responses:
        "201":
          description: "Issue Updated"
          schema:
            $ref: "#/definitions/Issue"
        "401":
          description: "Invalid"
          schema:
            $ref: "#/definitions/Error"
    get:
      summary: "retrieving Issue by Id"
      description: "This operation retrieves a Issue entity."
      tags:
        - "issue"
      parameters:
        - name: "issueid"
          description: "identification of the user who creates entity"
          required: true
          type: string
          in: path
      responses:
        "201":
          description: "Success"
          schema:
            $ref: "#/definitions/Issue"
        "401":
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
  /issue/:
    get:
      summary: "retrieving Issue by Id"
      description: "This operation retrieves a Issue entity."
      tags:
        - "issue"
      responses:
        "201":
          description: "Success"
          schema:
            $ref: "#/definitions/ArrayOfIssues"
        "401":
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
definitions:
  Event_Create:
    type: "object"
    description: "Event is the created"
    required:
      - userId
    properties:
      user_id:
        type: "string"
        description: "user id with which creates the event"
      name:
        type: "string"
        description: "name of the event"
      time_of_event:
        type: "string"
        description: "time of event"
      short_description:
        type: "string"
        description: "short description of event"
      summary:
        type: "string"
        description: "long summary of the event"
      event_location:
        $ref: "#/definitions/Location"
      country:
        type: "string"
        description: "country of the event"
      status:
        type: "string"
        description: "status of event"
        enum: [ "created","active","postponed","terminated","completed" ]
  Location_Ref:
    type: "object"
    description: "location of the event's place or origin"
    properties:
      location_id:
        type: "string"
      latitude:
        type: "string"
        description: "latitude of home_location"
      longitude:
        type: "string"
        description: "longitude of home_location"
      address:
        type: "string"
        description: "location of user entered by user manually"
  Location:
    type: "object"
    description: "location of the event's place or origin"
    properties:
      latitude:
        type: "string"
        description: "latitude of home_location"
      longitude:
        type: "string"
        description: "longitude of home_location"
      address:
        type: "string"
        description: "location of user entered by user manually"
  Event:
    type: "object"
    description: "returned event details"
    properties:
      user_id:
        type: "string"
        description: "user id with which created the event"
      event_id:
        type: "string"
        description: "event id of created event"
      name:
        type: "string"
        description: "name of the event"
      time_of_event:
        type: "string"
        description: "time of event"
      short_description:
        type: "string"
        description: "short description of event"
      summary:
        type: "string"
        description: "long summary of the event"
      event_location:
        $ref: "#/definitions/Location_Ref"
      country:
        type: "string"
        description: "country of the event"
      support:
        type: "integer"
        description: "number of votes"
      status:
        type: "string"
        description: "status of event"
        enum: [ "created","active","postponed","terminated","completed" ]
  ArrayOfEvents:
    type: "array"
    description: "return all events"
    items:
      $ref: "#/definitions/Event"
  Issue_Create:
    type: "object"
    description: "Issue is the created"
    required:
      - userId
    properties:
      user_id:
        type: "string"
        description: "user id with which creates the issue"
      name:
        type: "string"
        description: "name of the issue"
      time_of_Issue:
        type: "string"
        description: "time of issue"
      short_description:
        type: "string"
        description: "short description of issue"
      summary:
        type: "string"
        description: "long summary of the issue"
      event_location:
        $ref: "#/definitions/Location"
      country:
        type: "string"
        description: "country of the Issue"
      status:
        type: "string"
        description: "status of Issue"
        enum: [ "created","active","postponed","terminated","completed" ]
  Issue:
    type: "object"
    description: "returned issues details"
    required:
      - user_id
      - event_id
    properties:
      user_id:
        type: "string"
        description: "user id with which created the issues"
      event_id:
        type: "string"
        description: "event id of created issues"
      name:
        type: "string"
        description: "name of the issues"
      time_of_event:
        type: "string"
        description: "time of issues"
      short_description:
        type: "string"
        description: "short description of issues"
      summary:
        type: "string"
        description: "long summary of the issues"
      event_location:
        $ref: "#/definitions/Location_Ref"
      country:
        type: "string"
        description: "country of the issues"
      support:
        type: "integer"
        description: "number of votes"
      status:
        type: "string"
        description: "status of issues"
        enum: [ "created","active","postponed","terminated","completed" ]
  ArrayOfIssues:
    type: "array"
    description: "return all issues"
    items:
      $ref: "#/definitions/Issue"
  Error:
    type: "object"
    description: "error codes"
    properties:
      code:
        type: "string"
        description: "returned code"
      message:
        type: "string"
        description: "returned error message"
      status:
        type: "string"
        description: "returned status"
      reason:
        type: "string"
        description: "returned reason"